refactor: Consul Catalog labels.
